+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Ability to Block-list plugins that should not be installed on the jenkins dogu
-  - Block-listed plugins will be automatically removed when found to be installed
+  - Block-listed plugins will be automatically removed when found to be installed on dogu start and restart
 
 ## [v2.452.4-2] - 2024-09-18
 ### Changed