+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Upgrade Jenkins to 2.462.2; #171
+- Upgrade ces-build-lib to version 2.4.0
+- Upgrade dogu-build-lib to version 2.5.0
 
 ## [v2.452.4-2] - 2024-09-18
 ### Changed